Bug fixes and external texture support (#554)

+++ Assembly-CSharp/Global/Event/Engine/EventEngine.DoEventCode.cs
@@ -2683,7 +2683,6 @@ public partial class EventEngine
                 this.fieldmap.walkMesh.BGI_animShowFrame((UInt32)this.getv1(), (UInt32)this.getv1());
                 return 0;
             }
-            // TODO: no more information from Hades Workshop
             case EBin.event_code_binary.PLAYER_EQUIP: // "SetCharacterEquipment" Change the piece of equipment of a player, using it from the player's inventory
             {
                 CharacterId charId = this.chr2slot(this.getv3()); // character to (re-)equip.
@@ -2788,11 +2787,11 @@ public partial class EventEngine
                 }
                 return 1;
             }
-            case EBin.event_code_binary.TURN_OBJ_EX:
+            case EBin.event_code_binary.TURN_OBJ_EX: // "TurnTowardObjectEx" Make the specified character turn toward the specified object
             {
-                Actor turner = this.GetObj3() as Actor;
-                PosObj target = this.GetObj3() as PosObj;
-                Int32 tspeed = this.getv3();
+                Actor turner = this.GetObj3() as Actor; // character to turn
+                PosObj target = this.GetObj3() as PosObj; // object to look at
+                Int32 tspeed = this.getv3(); // turn speed
                 if (turner != null && target != null)
                 {
                     Single angle = this.eBin.angleAsm(target.pos[0] - turner.pos[0], target.pos[2] - turner.pos[2]);
@@ -2800,11 +2799,11 @@ public partial class EventEngine
                 }
                 return 0;
             }
-            case EBin.event_code_binary.AANIM_EX:
+            case EBin.event_code_binary.AANIM_EX: // "SetLogicalAnimationEx" Change the logical animation of the specified entry
             {
-                actor = this.GetObj3() as Actor;
-                Int32 kind = this.getv3();
-                UInt16 anim = (UInt16)(Int32)this.getv3();
+                actor = this.GetObj3() as Actor; // object re-animated
+                Int32 kind = this.getv3(); // type of logical animation
+                UInt16 anim = (UInt16)this.getv3(); // animation to use
                 switch (kind)
                 {
                     case 0: actor.idle = anim; break;
@@ -2818,21 +2817,21 @@ public partial class EventEngine
                 AnimationFactory.AddAnimWithAnimatioName(actor.go, FF9DBAll.AnimationDB.GetValue((Int32)anim));
                 return 0;
             }
-            case EBin.event_code_binary.VECTOR_CLEAR:
+            case EBin.event_code_binary.VECTOR_CLEAR: // "ClearMemoriaVector" Empty a vector in the vector system provided by Memoria
             {
-                Int32 vectID = this.getv3();
+                Int32 vectID = this.getv3(); // vector ID to clear
                 if (FF9StateSystem.EventState.gScriptVector.TryGetValue(vectID, out List<Int32> vect))
                     vect.Clear();
                 return 0;
             }
-            case EBin.event_code_binary.DICTIONARY_CLEAR:
+            case EBin.event_code_binary.DICTIONARY_CLEAR: // "ClearMemoriaDictionary" Empty a dictionary in the dictionary system provided by Memoria
             {
-                Int32 dictID = this.getv3();
+                Int32 dictID = this.getv3(); // dictionary ID to clear
                 if (FF9StateSystem.EventState.gScriptDictionary.TryGetValue(dictID, out Dictionary<Int32, Int32> dict))
                     dict.Clear();
                 return 0;
             }
-            case EBin.event_code_binary.BGLMOVE_TIMED:
+            case EBin.event_code_binary.BGLMOVE_TIMED: // "SetTilePositionTimed" Move a tile block over time
             {
                 this.fieldmap.EBG_overlayMoveTimed(this.getv3(), this.getv3(), this.getv3(), this.getv3(), this.getv3());
                 return 0;
